fix(api): accept entity identifier form in URLs + int64 id consistency (#172, #173)
Task/incident/corrective-action/audit-finding/legal endpoints parsed :id with raw
strconv.Atoi, so only a numeric id worked and the identifier form (TASK-6, …) 400'd.
Route them through parseID (extended to strip every NextIdentifier prefix incl.
legacy AST-), and standardize those entities' id to int64 to match suppliers/
assets/risks/systems — so parseID drops in and the whole API is on one id
vocabulary.

Also fixes the AST-/ASSET- split: the suggestion-apply path now uses the shared
nextIdentifierTx(…, "asset") helper (ASSET-) instead of a hardcoded AST-.

Regression test asserts identifier and numeric URLs resolve the same record for
tasks/incidents/CAs/legal (get/put/delete). go build + go unit tests green.

Closes #172. Closes #173.

+++ internal/isms/api/api_readings.go
@@ -351,7 +351,7 @@ func (s *Server) handleCreateLegalReading(c echo.Context) error {
 // Each reading just updates current likelihood/impact and recomputes score.
 // nextReview, if provided, sets the legal requirement's next review date (YYYY-MM-DD).
 func writeLegalFromReading(ctx context.Context, tx pgx.Tx, s *Server, orgID int, legalID int, r *db.EntityReading, actor string, nextReview string) error {
-	lr, err := s.db.GetLegalRequirement(ctx, orgID, legalID)
+	lr, err := s.db.GetLegalRequirement(ctx, orgID, int64(legalID))
 	if err != nil {
 		return fmt.Errorf("legal requirement %d not found: %w", legalID, err)
 	}
